### Refactor database handling and reduce redundancy
Updated several database specs and schema definitions to enhance clarity and reduce redundancy:
- Corrected table name for PostgreSQL schema queries.
- Simplified customer insertion logic by omitting redundant ID assignments.
- Adjusted column naming conventions for coherence across tests.
- Refined initialization of `CustomerModel` with optional IDs and replaced `Int64` balance with `Int32`.
- Consolidated query calls by directly referencing `Northwind.query`.
- Fixed group and index size expectations in schema tests.
- Added logging for SQL queries in the `first` method for better debugging.
- Split and executed SQL statements individually in schema transactions to ensure atomicity and prevent partial executions.

These changes improve maintainability and ensure more robust and readable tests, while facilitating easier debugging and future enhancements.

+++ spec/schemas/northwind.cr
@@ -7,7 +7,7 @@ Northwind = Cql::Schema.build(
     primary :id, Int32
     column :name, String
     column :city, String
-    column :balance, Int64
+    column :balance, Int32
     timestamps
   end
